change donkey_thrown_offset return type

+++ src/cpp.rs
@@ -13210,7 +13210,7 @@ pub mod root {
 						arg1: *mut root::app::FighterParamAccessor2,
 						arg2: libc::c_int,
 						arg3: libc::c_int,
-					) -> u64;
+					) -> f32;
 				}
 				extern "C" {
 					#[link_name = "\u{1}_ZN3app8lua_bind49FighterParamAccessor2__ridley_dragged_offset_implEPNS_21FighterParamAccessor2Eii"]
